fix(accessibility): refacto moodItem
